+++ packages/pharos/src/components/checkbox/pharos-checkbox.ts
@@ -2,7 +2,6 @@ import { html, property, query } from 'lit-element';
 import type { TemplateResult, CSSResultArray } from 'lit-element';
 import { ifDefined } from 'lit-html/directives/if-defined.js';
 import { checkboxStyles } from './pharos-checkbox.css';
-import { designTokens } from '../../styles/variables.css';
 import { unsafeSVG } from 'lit-html/directives/unsafe-svg.js';
 import { classMap } from 'lit-html/directives/class-map.js';
 import { customElement } from '../../utils/decorators';
@@ -15,7 +14,6 @@ import { FormElement } from '../base/form-element';
 import FormMixin from '../../utils/mixins/form';
 
 const LINKS = `a[href],pharos-link[href]`;
-const matchesFunc = 'matches' in Element.prototype ? 'matches' : 'msMatchesSelector';
 
 /**
  * Pharos checkbox component.
@@ -61,7 +59,7 @@ export class PharosCheckbox extends FormMixin(FormElement) {
   private _checkbox!: HTMLInputElement;
 
   public static get styles(): CSSResultArray {
-    return [designTokens, super.styles, checkboxStyles];
+    return [super.styles, checkboxStyles];
   }
 
   protected firstUpdated(): void {
@@ -92,7 +90,7 @@ export class PharosCheckbox extends FormMixin(FormElement) {
   }
 
   private _handleClick(event: Event): void {
-    if (!(event.target as Element)[matchesFunc](LINKS)) {
+    if (!(event.target as Element).matches(LINKS)) {
       event.preventDefault();
       event.stopPropagation();
       this._checkbox.click();
@@ -109,7 +107,7 @@ export class PharosCheckbox extends FormMixin(FormElement) {
   }
 
   private _handleMousedown(event: MouseEvent): void {
-    if (!(event.target as Element)[matchesFunc](LINKS)) {
+    if (!(event.target as Element).matches(LINKS)) {
       event.preventDefault();
     }
   }