+++ packages/pharos/src/components/dropdown-menu/pharos-dropdown-menu.ts
@@ -5,7 +5,6 @@ import { createPopper } from '../../utils/popper';
 import debounce from '../../utils/debounce';
 import observeResize from '../../utils/observeResize';
 import { dropdownMenuStyles } from './pharos-dropdown-menu.css';
-import { designTokens } from '../../styles/variables.css';
 import { customElement } from '../../utils/decorators';
 import type { PharosDropdownMenuItem } from './pharos-dropdown-menu-item';
 import type { PharosDropdownMenuNavLink } from '../dropdown-menu-nav/pharos-dropdown-menu-nav-link';
@@ -88,7 +87,7 @@ export class PharosDropdownMenu extends FocusMixin(OverlayElement) {
   }
 
   public static get styles(): CSSResultArray {
-    return [designTokens, dropdownMenuStyles];
+    return [dropdownMenuStyles];
   }
 
   public removeAllTriggers(): void {