+++ packages/pharos/src/components/tabs/pharos-tabs.ts
@@ -1,12 +1,9 @@
 import { html, LitElement } from 'lit-element';
 import type { TemplateResult, CSSResultArray } from 'lit-element';
 import { tabsStyles } from './pharos-tabs.css';
-import { designTokens } from '../../styles/variables.css';
 import { customElement } from '../../utils/decorators';
 import type { PharosTab } from './pharos-tab';
 import type { PharosTabPanel } from './pharos-tab-panel';
-
-const matchesFunc = 'matches' in Element.prototype ? 'matches' : 'msMatchesSelector';
 
 /**
  * Pharos tabs component.
@@ -20,7 +17,7 @@ const matchesFunc = 'matches' in Element.prototype ? 'matches' : 'msMatchesSelec
 @customElement('pharos-tabs')
 export class PharosTabs extends LitElement {
   public static get styles(): CSSResultArray {
-    return [designTokens, tabsStyles];
+    return [tabsStyles];
   }
 
   protected firstUpdated(): void {
@@ -101,7 +98,7 @@ export class PharosTabs extends LitElement {
     const ids = tabs.map((tab) => tab.id);
 
     const focused = document.activeElement as PharosTab;
-    if (!focused[matchesFunc]('pharos-tab')) {
+    if (!focused.matches('pharos-tab')) {
       return;
     }
 
@@ -127,7 +124,7 @@ export class PharosTabs extends LitElement {
   }
 
   private _handleFocusout(event: FocusEvent): void {
-    if (event.relatedTarget && (event.relatedTarget as Element)[matchesFunc]('pharos-tab')) {
+    if (event.relatedTarget && (event.relatedTarget as Element).matches('pharos-tab')) {
       return;
     }
     const tabs = this.querySelectorAll(`pharos-tab`) as NodeListOf<PharosTab>;