+++ packages/pharos/tasks/build-styles.js
@@ -23,13 +23,7 @@ const buildStyles = async () => {
     // Convert to CSS
     .pipe(sass())
     // Run through PostCSS
-    .pipe(
-      postcss([
-        autoprefixer({
-          grid: 'autoplace',
-        }),
-      ])
-    )
+    .pipe(postcss([autoprefixer()]))
     // Wrap in lit-element css template tag
     .pipe(
       through2.obj((file, enc, done) => {
@@ -72,13 +66,7 @@ const buildSlotStyles = async () => {
     // Convert to CSS
     .pipe(sass())
     // Run through PostCSS
-    .pipe(
-      postcss([
-        autoprefixer({
-          grid: 'autoplace',
-        }),
-      ])
-    )
+    .pipe(postcss([autoprefixer()]))
     // Update filename
     .pipe(
       rename((path) => {